get pattern

+++ server/models/main.py
@@ -18,13 +18,12 @@ def suggested_guesses(k):
     
     prev_guesses, patterns, possibilities = util.get_guesses_patterns_possibilities()
     choices = all_words # TODO: always true?
-    
+    print("Previous guesses: ", prev_guesses)
     if prev_guesses != []: # If not beginning of the game
-        print(prev_guesses)
         possibilities = util.get_possible_words(prev_guesses[-1], patterns[-1], possibilities)
         util.update_possibilities(possibilities)
     print(len(possibilities), "possibilities now")
-    # print(possibilities)
+    print(possibilities)
 
     print("Optimal ", k, " choices: ", opt.guess(choices, possibilities, priors, k))
 
@@ -39,7 +38,7 @@ def update_game_state(guess, pattern):
 
 def update(answer, guesses):
     if guesses:
-        update_game_state(guesses[-1], 0) # TODO: WRITE A FUNCTION TO FIGURE OUT PATTERN BASED ON WORD
+        update_game_state(guesses[-1], util.get_pattern(guesses[-1], answer))
         suggested_guesses(10)
     else: # first guess
         reset()
@@ -47,31 +46,37 @@ def update(answer, guesses):
 
 def reset():
     """ When the game is done, erase files storing game state"""
-    print("resetting")
-    os.remove(Config.PATTERN_GRID_DATA)
-    os.remove(Config.GUESSES_FILE)
-    os.remove(Config.PATTERNS_FILE)
-    os.remove(Config.POSSIBILITIES_FILE)
+    if os.path.exists(Config.PATTERN_GRID_DATA):
+        os.remove(Config.PATTERN_GRID_DATA)
+    if os.path.exists(Config.GUESSES_FILE):
+        os.remove(Config.GUESSES_FILE)
+    if os.path.exists(Config.PATTERNS_FILE):
+        os.remove(Config.PATTERNS_FILE)
+    if os.path.exists(Config.POSSIBILITIES_FILE):
+        os.remove(Config.POSSIBILITIES_FILE)
 
 
 
 """
 A pattern for two words represents the wordle-similarity
-pattern (grey -> 0, yellow -> 1, green -> 2) but as an integer
+pattern (grey -> 0, yellow -> 2, green -> 1) but as an integer
 between 0 and 3^5. Reading this integer in ternary gives the
 associated pattern. Ie convert to base 3
 """
 # Example: answer is "where"
 if __name__ == "__main__":
-    # answer = sys.argv[1]
-    # guesses = sys.argv[2]
-    # if guesses == 'none': # parsed as none for the first guess to ensure there's still detectable content
-    #     guesses = []
-    # else:
-    #     guesses = guesses.split(",")
-    # update(answer, guesses)
-    suggested_guesses(10)
-    update_game_state("raise", 83) # pattern is 10002
-    suggested_guesses(10)
-    update_game_state("would", 162) # pattern is 20000
-    suggested_guesses(10)
+    answer = sys.argv[1]
+    guesses = sys.argv[2]
+    if guesses == 'none': # parsed as none for the first guess to ensure there's still detectable content
+        guesses = []
+    else:
+        guesses = guesses.split(",")
+    update(answer, guesses)
+
+
+    # suggested_guesses(10)
+    # update_game_state("raise", 83) # pattern is 10002
+    # suggested_guesses(10)
+    # update_game_state("would", 162) # pattern is 20000
+    # suggested_guesses(10)
+    # reset()
